bug fix #002:correct a bug that make progressbar still,because of type cast error in putExtra

+++ app/src/main/java/com/busysnail/filedownload/db/ThreadDAO.java
@@ -28,7 +28,7 @@ public interface ThreadDAO {
      * @param thread_id
      * @param finishded
      */
-    void updateThread(String url,int thread_id,int finishded);
+    void updateThread(String url,int thread_id,long finishded);
 
     /**
      * 查询文件的线程信息
